@@ -1,6 +1,6 @@
 {
   "name": "@wearenova/mui-data-table",
-  "version": "1.6.0",
+  "version": "2.0.0-rc.1",
   "description": "An extensive data-table with sorting, multi-column filtering, pagination and more with server-side data fetching support.",
   "source": "src/index.tsx",
   "types": "lib/index.d.ts",
@@ -54,12 +54,11 @@
     ]
   },
   "peerDependencies": {
-    "@material-ui/core": "^4.0.0",
-    "@material-ui/icons": "^4.0.0",
-    "@material-ui/pickers": "^3.3.0",
-    "@material-ui/styles": "^4.0.0",
-    "react": "^16.8.0 || ^17.0.0",
-    "react-dom": "^16.8.0 || ^17.0.0"
+    "@mui/icons-material": "^5.0.0",
+    "@mui/lab": "^5.0.0-alpha.30",
+    "@mui/material": "^5.0.0",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0"
   },
   "dependencies": {
     "dot-prop": "^6.0.1",
