@@ -1,16 +1,22 @@
-import { Divider, IconButton, makeStyles, TableSortLabel, Tooltip } from "@material-ui/core";
-import AcUnit from "@material-ui/icons/AcUnit";
-import FilterList from "@material-ui/icons/FilterList";
-import Visibility from "@material-ui/icons/Visibility";
-import { createStyles } from "@material-ui/styles";
-import clsx from "clsx";
+import AcUnit from "@mui/icons-material/AcUnit";
+import FilterList from "@mui/icons-material/FilterList";
+import Visibility from "@mui/icons-material/Visibility";
+import { Box, CSSObject, Divider, IconButton, styled, TableSortLabel, Tooltip } from "@mui/material";
 import PropTypes from "prop-types";
 import React, { Fragment, MouseEventHandler, PropsWithChildren, useCallback, useContext, useMemo, useRef } from "react";
 import { InitialFilterValues } from "./Filter";
 import TableContext, { TableState } from "./table.context";
 import type { ActionButton, BaseData, ColGroupDefinition, ColumnDefinition, Sort } from "./table.types";
+import { TableCellAlign } from "./table.types";
 import TableCell from "./TableCell.component";
-import { dispatchTableEvent, getColumnTitle, getDataType, getDefaultOperator, getPath } from "./utils";
+import {
+  dispatchTableEvent,
+  dontForwardProps,
+  getColumnTitle,
+  getDataType,
+  getDefaultOperator,
+  getPath,
+} from "./utils";
 import { ColumnDefinitionPropType } from "./_propTypes";
 
 interface HeaderCellProps<RowType extends BaseData, AllDataType extends RowType[]> {
@@ -23,107 +29,97 @@ interface HeaderCellProps<RowType extends BaseData, AllDataType extends RowType[
   style?: React.CSSProperties;
 }
 
-const useStyles = makeStyles(
-  (theme) =>
-    createStyles({
-      tableCell: {
-        "&:hover $filterIconButton": {
-          opacity: 1,
-        },
-      },
-      alignLeft: {
-        justifyContent: "flex-start",
-        "&$editableOffset": {
-          paddingLeft: theme.spacing(1),
-        },
-      },
-      alignCenter: {
-        justifyContent: "center",
-        "& > div": {
-          flex: 1,
-          width: "100%",
-          "&:nth-child(1)": {
-            order: 2,
-            textAlign: "center",
-          },
-          "&:nth-child(2)": {
-            order: 1,
-          },
-          "&:nth-child(3)": {
-            order: 3,
-          },
-        },
-      },
-      alignRight: {
-        justifyContent: "flex-end",
-        "&$editableOffset": {
-          padding: 0,
-          paddingRight: theme.spacing(1),
-        },
-        "& > div:nth-child(1)": {
-          order: 3,
-          textAlign: "right",
-          "&  $headerCellSortLabel": {
-            flexDirection: "row-reverse",
-          },
-        },
-        "& > div:nth-child(2)": {
-          order: 2,
-        },
-        "& > div:nth-child(3)": {
-          order: 1,
-        },
-      },
-      editableOffset: {},
-      filterIconButton: {
-        opacity: 0.2,
-        transition: theme.transitions.create(["opacity", "color"], {
-          duration: theme.transitions.duration.shorter,
-          easing: theme.transitions.easing.easeInOut,
-        }),
-      },
-      headerCellBody: {
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center",
-        whiteSpace: "pre",
-      },
-      headerCellButtonGroup: {
-        display: "flex",
-        alignItems: "center",
-        "& svg": {
-          fontSize: "1rem",
-        },
-      },
-      headerCellInner: {
-        display: "flex",
-        position: "relative",
-      },
-      headerCellSortLabel: {
-        whiteSpace: "inherit",
-        "& > svg.MuiTableSortLabel-icon": {
-          opacity: 0.2,
-        },
-      },
-      resizeable: {
-        paddingRight: theme.spacing(2),
-        overflow: "hidden",
-      },
-      resizeHandle: {
-        cursor: "col-resize",
-        position: "absolute",
-        right: 0,
-        width: 3,
-        "&:active,&:hover": {
-          backgroundColor: theme.palette.action.active,
-        },
-      },
-      stickyColGroupHeader: {
-        left: "unset",
-      },
+const FilterButton = styled(IconButton, {
+  label: "DataTable-FilterButton",
+  shouldForwardProp: dontForwardProps("active"),
+})<{ active: boolean }>(({ active, theme }) => [
+  {
+    opacity: 0.3,
+    transition: theme.transitions.create(["opacity", "color"], {
+      duration: theme.transitions.duration.shorter,
+      easing: theme.transitions.easing.easeInOut,
     }),
-  { name: "DataTable-HeaderCell" },
-);
+  },
+  active && {
+    opacity: 1,
+  },
+]);
+
+const HeaderTableCell = styled(TableCell, {
+  label: "DataTable-HeaderTableCell",
+  shouldForwardProp: dontForwardProps("colGroupHeader", "pinned"),
+})<{
+  colGroupHeader: boolean;
+  pinned: boolean;
+}>(({ colGroupHeader, pinned }) => [
+  {
+    [`&:hover ${FilterButton}`]: {
+      opacity: 1,
+    },
+  },
+  colGroupHeader && pinned && { left: "unset" },
+]);
+
+const alignmentStyles: Record<TableCellAlign, CSSObject> = {
+  left: {
+    justifyContent: "flex-start",
+  },
+  center: {
+    justifyContent: "center",
+    "& > div": {
+      flex: 1,
+      width: "100%",
+      "&:nth-of-type(1)": {
+        order: 2,
+        textAlign: "center",
+      },
+      "&:nth-of-type(2)": {
+        order: 1,
+      },
+      "&:nth-of-type(3)": {
+        order: 3,
+      },
+    },
+  },
+  right: {
+    justifyContent: "flex-end",
+    "& > div:nth-of-type(1)": {
+      order: 3,
+      textAlign: "right",
+      "& .MuiTableSortLabel-root": {
+        flexDirection: "row-reverse",
+      },
+    },
+    "& > div:nth-of-type(2)": {
+      order: 2,
+    },
+    "& > div:nth-of-type(3)": {
+      order: 1,
+    },
+  },
+};
+
+const editableOffsetKey: Record<Exclude<TableCellAlign, "center">, string> = {
+  left: "paddingLeft",
+  right: "paddingRight",
+};
+
+const InnerHeaderCell = styled("div", {
+  label: "DataTable-InnerHeaderCell",
+  shouldForwardProp: dontForwardProps("alignment", "editableOffset", "resizeable"),
+})<{
+  alignment?: TableCellAlign;
+  editableOffset: boolean;
+  resizeable: boolean;
+}>(({ alignment = "left", editableOffset, resizeable, theme }) => [
+  { display: "flex", position: "relative" },
+  resizeable && { paddingRight: theme.spacing(2), overflow: "hidden" },
+  alignment && alignmentStyles[alignment],
+  editableOffset &&
+    alignment !== "center" && {
+      [editableOffsetKey[alignment]]: theme.spacing(1),
+    },
+]);
 
 /**
  * The HeaderCell component is a wrapper around the custom TableCell component which manages all the state for a single header cell
@@ -138,11 +134,8 @@ const HeaderCell = <RowType extends BaseData, AllDataType extends RowType[] = Ro
   onFilterClick,
   hasColGroups = false,
   colGroupHeader = false,
-  className,
   style,
-  ...props
 }: PropsWithChildren<HeaderCellProps<RowType, AllDataType>>) => {
-  const classes = useStyles(props);
   const { activeFilters, sort, enableHiddenColumns, hiddenColumns, pinnedColumn, allTableData, update, resizeable } =
     useContext<TableState<RowType, AllDataType>>(TableContext);
   const tableCellRef = useRef<HTMLTableCellElement>(null);
@@ -206,14 +199,6 @@ const HeaderCell = <RowType extends BaseData, AllDataType extends RowType[] = Ro
     [id, update],
   );
 
-  const headerClasses = useMemo(
-    () =>
-      clsx(classes.tableCell, className, {
-        [classes.stickyColGroupHeader]: colGroupHeader && pinnedColumn !== id,
-      }),
-    [className, classes.stickyColGroupHeader, classes.tableCell, colGroupHeader, id, pinnedColumn],
-  );
-
   const handleFilterClick = useCallback<MouseEventHandler<HTMLButtonElement>>(
     (e) => {
       if (!filterPath) return;
@@ -234,30 +219,33 @@ const HeaderCell = <RowType extends BaseData, AllDataType extends RowType[] = Ro
   return (
     <Fragment key={id}>
       <Tooltip title={isHidden ? `Unhide '${structure.title}' Column` : ""} placement="top">
-        <TableCell
+        <HeaderTableCell
           onClick={handleUnhide}
           ref={tableCellRef}
           hidden={Boolean(isHidden)}
           pinned={pinnedColumn === id}
+          colGroupHeader={colGroupHeader}
           colSpan={colSpan}
           rowSpan={rowSpan}
           align={structure.align}
-          className={headerClasses}
           style={style}
         >
-          <div
-            className={clsx(classes.headerCellInner, classes.alignLeft, {
-              [classes.alignRight]: structure.align === "right",
-              [classes.alignCenter]: structure.align === "center",
-              [classes.editableOffset]: Boolean(structure.editable),
-              [classes.resizeable]: !colGroupHeader && resizeable,
-            })}
+          <InnerHeaderCell
+            alignment={structure.align}
+            resizeable={!colGroupHeader && resizeable}
+            editableOffset={Boolean(structure.editable)}
           >
-            <div className={classes.headerCellBody}>
+            <Box
+              sx={{
+                display: "flex",
+                alignItems: "center",
+                justifyContent: "center",
+                whiteSpace: "pre",
+              }}
+            >
               {structure.sorter ? (
                 <TableSortLabel
                   onClick={handleSort}
-                  className={classes.headerCellSortLabel}
                   active={
                     sort.direction &&
                     (sort.key === structure.sorter || sort.key === structure.dataIndex || sort.key === id)
@@ -267,14 +255,28 @@ const HeaderCell = <RowType extends BaseData, AllDataType extends RowType[] = Ro
                       ? sort.direction
                       : undefined
                   }
+                  sx={{
+                    whiteSpace: "inherit",
+                    "& > svg.MuiTableSortLabel-icon": {
+                      opacity: 0.2,
+                    },
+                  }}
                 >
                   {headerTitle}
                 </TableSortLabel>
               ) : (
                 headerTitle
               )}
-            </div>
-            <div className={classes.headerCellButtonGroup}>
+            </Box>
+            <Box
+              sx={{
+                display: "flex",
+                alignItems: "center",
+                "& svg": {
+                  fontSize: "1rem",
+                },
+              }}
+            >
               {enableHiddenColumns && (
                 <IconButton onClick={handleHiddenColumnsChange} size="small">
                   <Visibility />
@@ -291,25 +293,37 @@ const HeaderCell = <RowType extends BaseData, AllDataType extends RowType[] = Ro
                 </IconButton>
               ))}
               {filterEnabled && (
-                <IconButton
+                <FilterButton
                   onClick={handleFilterClick}
                   onMouseUp={stopPropagation}
                   onTouchEnd={stopPropagation}
                   data-testid="tableFilterButton"
                   color={filterActive ? "primary" : "default"}
+                  active={filterActive}
                   size="small"
-                  className={clsx({ [classes.filterIconButton]: !filterActive })}
                 >
                   <FilterList />
-                </IconButton>
+                </FilterButton>
               )}
-            </div>
+            </Box>
             <div />
             {!colGroupHeader && resizeable && (
-              <Divider id="DataTable-ResizeHandle" className={classes.resizeHandle} orientation="vertical" />
+              <Divider
+                id="DataTable-ResizeHandle"
+                orientation="vertical"
+                sx={{
+                  cursor: "col-resize",
+                  position: "absolute",
+                  right: 0,
+                  borderRightWidth: 3,
+                  "&:active,&:hover": {
+                    borderColor: "action.active",
+                  },
+                }}
+              />
             )}
-          </div>
-        </TableCell>
+          </InnerHeaderCell>
+        </HeaderTableCell>
       </Tooltip>
     </Fragment>
   );
