@@ -34,6 +34,10 @@ let defaultCurrency = "GBP";
 
 const TABLE_EVENTS = ["*", "cancelEdit", "closeFilter"] as const;
 
+export function dontForwardProps(...propNamesToIgnore: string[]) {
+  return (name: string) => propNamesToIgnore.indexOf(name) === -1;
+}
+
 /**
  * A function to dispatch a custom Data Table event.
  *
@@ -208,7 +212,6 @@ const OPERATOR_MAP: Record<OperatorValues, (args: MatchActionArg) => boolean> = 
  * @param filter.value the filter value
  * @param filter.operator the filter operator
  * @param filter.type the filter type
- * @param utils the date utilities from \@material-ui/pickers
  * @returns whether the value matches the filter value given the filter operator and type
  */
 export function getMatch<RowType extends BaseData>(value: RowType[keyof RowType], filter: ActiveFilter): boolean {
@@ -223,7 +226,6 @@ export function getMatch<RowType extends BaseData>(value: RowType[keyof RowType]
  *
  * @param data table data
  * @param filters the filters to apply
- * @param utils the date utilities from \@material-ui/pickers
  * @returns the filtered data
  */
 export function getFilteredData<RowType extends BaseData>(data: RowType[], filters: ActiveFilters<RowType>): RowType[] {
